+++ ThirdPartyAdapters/applovin/applovin/src/main/java/com/applovin/mediation/rtb/AppLovinRtbBannerRenderer.java
@@ -1,7 +1,7 @@
 package com.applovin.mediation.rtb;
 
 import android.content.Context;
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 import android.util.Log;
 import android.view.View;
 import com.applovin.adview.AppLovinAdView;