+++ ThirdPartyAdapters/maio/maio/src/main/java/jp/maio/sdk/android/mediation/admob/adapter/MaioAdsManager.java
@@ -1,7 +1,7 @@
 package jp.maio.sdk.android.mediation.admob.adapter;
 
 import android.app.Activity;
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 import android.text.TextUtils;
 import android.util.Log;